ai-sentiment-analysis/requirements.txt
# Reddit API
praw>=7.7.1

# Data
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0

# NLP & Sentiment
transformers>=4.35.0
torch>=2.1.0
vaderSentiment>=3.3.2
nltk>=3.8.1

# Topic Modeling
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
wordcloud>=1.9.2

# Dashboard
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
